+++ sklearn/cluster/__init__.py
@@ -3,18 +3,18 @@ The :mod:`sklearn.cluster` module gathers popular unsupervised clustering
 algorithms.
 """
 
-from .spectral import spectral_clustering, SpectralClustering
-from .mean_shift_ import (mean_shift, MeanShift,
+from ._spectral import spectral_clustering, SpectralClustering
+from ._mean_shift import (mean_shift, MeanShift,
                           estimate_bandwidth, get_bin_seeds)
 from ._affinity_propagation import affinity_propagation, AffinityPropagation
 from ._hierarchical import (ward_tree, AgglomerativeClustering, linkage_tree,
                             FeatureAgglomeration)
 from ._k_means import k_means, KMeans, MiniBatchKMeans
 from ._dbscan import dbscan, DBSCAN
-from .optics_ import (OPTICS, cluster_optics_dbscan, compute_optics_graph,
+from ._optics import (OPTICS, cluster_optics_dbscan, compute_optics_graph,
                       cluster_optics_xi)
 from ._bicluster import SpectralBiclustering, SpectralCoclustering
-from .birch import Birch
+from ._birch import Birch
 
 __all__ = ['AffinityPropagation',
            'AgglomerativeClustering',
