+++ sklearn/cluster/tests/test_optics.py
@@ -6,7 +6,7 @@ import numpy as np
 import pytest
 
 from sklearn.datasets.samples_generator import make_blobs
-from sklearn.cluster.optics_ import (OPTICS,
+from sklearn.cluster._optics import (OPTICS,
                                      _extend_region,
                                      _extract_xi_labels)
 from sklearn.metrics.cluster import contingency_matrix
